agregado room y dependencias. refactor de algunas cosas de retrofit
