+++ app/src/main/java/com/example/parcialtp3/ApiInterface/ApiBuilder.kt
@@ -2,17 +2,19 @@ package com.example.parcialtp3.ApiInterface
 
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
+import retrofit2.create
 
-object DogRetrofit {
+object ApiBuilder {
 
     private val URL_BASE_API  = "https://dog.ceo/api/breed/"
 
-    private fun getRetrofit(): DogCeoApi {
-        return Retrofit.Builder()
-            .baseUrl(URL_BASE_API)
-            .addConverterFactory(GsonConverterFactory.create())
-            .build()
-            .create(DogCeoApi::class.java)
+    private val retrofit = Retrofit.Builder()
+        .baseUrl(URL_BASE_API)
+        .addConverterFactory(GsonConverterFactory.create())
+        .build()
+
+    fun create(): DogCeoApi {
+        return retrofit.create(DogCeoApi::class.java)
     }
-    
+
 }
