+++ app/src/main/java/com/example/parcialtp3/ApiInterface/DogCeoApi.kt
@@ -1,15 +1,12 @@
 package com.example.parcialtp3.ApiInterface
 
-import com.example.parcialtp3.DogModel.Dog
-import retrofit2.Call
+import com.example.parcialtp3.DataModels.DogResponse
 import retrofit2.Response
 import retrofit2.http.GET
 import retrofit2.http.Url
 
-interface ApiInterface {
+interface DogCeoApi {
     @GET
-    suspend fun getDogsByBreeds(@Url url:String):Response<Dog>
-
-    suspend fun getBreeds(@Url url:String):Response
+    suspend fun getDogsByBreeds(@Url url:String):Response<DogResponse>
 
 }
