+++ app/src/main/java/com/example/parcialtp3/DataModels/DogResponse.kt
@@ -1,4 +1,4 @@
-package com.example.parcialtp3.Models
+package com.example.parcialtp3.DataModels
 
 import com.google.gson.annotations.SerializedName
